refactor: 필터정렬 로직 재수정

+++ core/database/src/main/java/com/into/websoso/core/database/entity/InDatabaseFilteredNovelEntity.kt
@@ -8,7 +8,7 @@ data class InDatabaseFilteredNovelEntity(
     @PrimaryKey val userNovelId: Long,
     val novelId: Long,
     val title: String,
-    val sortIndex: Long,
+    val sortIndex: Int,
     val novelImage: String,
     val novelRating: Float,
     val readStatus: String,
